Adding support for new 2 connection types: "ADPS2" and "MQTTBroker". Adding unit tests for the new connection types data. Modifying ADUC_JSON_GetBooleanField to return result in an out param instead.

+++ src/utils/config_utils/inc/aduc/config_utils.h
@@ -127,7 +127,7 @@ typedef struct tagADUC_ConfigInfo
 
     ADUC_AgentInfo* agents; /**< Array of agents that are configured. */
 
-    size_t agentCount; /**< Total number of agents configured. */
+    unsigned int agentCount; /**< Total number of agents configured. */
 
     const char* compatPropertyNames; /**< Compat property names. */
